@@ -12,7 +12,7 @@
   ],
   "license": "MIT",
   "require": {
-    "php": ">=7.3",
+    "php": ">=7.4",
     "ext-json": "*",
     "ext-pdo": "*",
     "setbased/exception": "^2.1",
