@@ -21,35 +21,35 @@ class SqlitePdoRoutineLoaderWorker extends SqlitePdoWorker implements RoutineLoa
    *
    * @var array
    */
-  private $errorFilenames = [];
+  private array $errorFilenames = [];
 
   /**
    * Class name for mangling routine and parameter names.
    *
-   * @var string
+   * @var string|null
    */
-  private $nameMangler;
+  private ?string $nameMangler;
 
   /**
    * The metadata of all stored routines. Note: this data is stored in the metadata file and is generated by PhpStratum.
    *
    * @var array
    */
-  private $phpStratumMetadata = [];
+  private array $phpStratumMetadata = [];
 
   /**
    * The filename of the file with the metadata of all stored routines.
    *
    * @var string
    */
-  private $phpStratumMetadataFilename;
+  private string $phpStratumMetadataFilename;
 
   /**
    * Pattern where of the sources files.
    *
    * @var string
    */
-  private $sourcePattern;
+  private string $sourcePattern;
 
   /**
    * All sources with stored routines. Each element is an array with the following keys:
@@ -61,7 +61,7 @@ class SqlitePdoRoutineLoaderWorker extends SqlitePdoWorker implements RoutineLoa
    *
    * @var array[]
    */
-  private $sources = [];
+  private array $sources = [];
 
   //--------------------------------------------------------------------------------------------------------------------
   /**
