@@ -20,105 +20,105 @@ class RoutineLoaderHelper
    *
    * @var string
    */
-  private $designationType;
+  private string $designationType;
 
   /**
    * An in memory SQLite database.
    *
    * @var SqlitePdoDataLayer
    */
-  private $dl;
+  private SqlitePdoDataLayer $dl;
 
   /**
    * The reflection of the DocBlock of the stored routine.
    *
-   * @var DocBlockReflection|null
+   * @var DocBlockReflection
    */
-  private $docBlockReflection;
+  private DocBlockReflection $docBlockReflection;
 
   /**
    * The Output decorator
    *
    * @var StratumStyle
    */
-  private $io;
+  private StratumStyle $io;
 
   /**
    * The offset of the first line of the payload of the stored routine ins the source file.
    *
    * @var int
    */
-  private $offset;
+  private int $offset;
 
   /**
    * The metadata of the stored routine. Note: this data is stored in the metadata file and is generated by PhpStratum.
    *
    * @var array
    */
-  private $phpStratumMetadata;
+  private array $phpStratumMetadata;
 
   /**
    * The replace pairs (i.e. placeholders and their actual values, see strst).
    *
    * @var array
    */
-  private $replace = [];
+  private array $replace = [];
 
   /**
    * A map from placeholders to their actual values.
    *
    * @var array
    */
-  private $replacePairs = [];
+  private array $replacePairs = [];
 
   /**
    * The return type of the stored routine (only if designation type singleton0, singleton1, or function).
    *
    * @var string|null
    */
-  private $returnType;
+  private ?string $returnType = null;
 
   /**
    * The name of the stored routine.
    *
    * @var string
    */
-  private $routineName;
+  private string $routineName;
 
   /**
    * The routine parameters.
    *
    * @var RoutineParametersHelper|null
    */
-  private $routineParameters = null;
+  private ?RoutineParametersHelper $routineParameters = null;
 
   /**
    * The payload of the stored routine (i.e. the code without the DocBlock).
    *
    * @var string
    */
-  private $routinePayLoad;
+  private string $routinePayLoad;
 
   /**
    * The source code as a single string of the stored routine.
    *
    * @var string
    */
-  private $routineSourceCode;
+  private string $routineSourceCode;
 
   /**
    * The source code as an array of lines string of the stored routine.
    *
    * @var array
    */
-  private $routineSourceCodeLines;
+  private array $routineSourceCodeLines;
 
   /**
    * The source filename holding the stored routine.
    *
    * @var string
    */
-  private $sourceFilename;
+  private string $sourceFilename;
 
   //--------------------------------------------------------------------------------------------------------------------
   /**
